Add PageExtended so no break of current Page
