+++ Microsoft.Azure.CosmosRepository/src/Paging/IPageExtended.cs
@@ -10,32 +10,36 @@ namespace Microsoft.Azure.CosmosRepository.Paging
     /// Represents a page of data from a cosmos query
     /// </summary>
     /// <typeparam name="T"></typeparam>
-    public interface IPage<out T> where T : IItem
+    public interface IPageExtended<out T>:IPage<T> where T : IItem
     {
+        /// <summary>
+        /// The total amount pages that matched the query.
+        /// </summary>
+        int TotalPages { get; }
+
         /// <summary>
         /// The total amount items that matched the query.
         /// </summary>
-        int Total { get; }
+        int? PageNumber { get; }
 
         /// <summary>
-        /// The size of the current page.
+        /// Gets a value indicating whether the are next pages.
         /// </summary>
-        int Size { get; }
+        bool HasNextPage { get; }
 
         /// <summary>
-        /// The items that are in the current page.
+        /// Gets a value indicating whether the are previous pages.
         /// </summary>
-        IReadOnlyList<T> Items { get; }
+        bool HasPreviousPage { get; }
 
         /// <summary>
-        /// The amount of RU's the given query cost.
+        /// The previous page number.
         /// </summary>
-        double Charge { get; }
+        int PreviousPageNumber { get; }
 
         /// <summary>
-        /// The continuation token used to load results from a stateless marker.
+        /// The next page number.
         /// </summary>
-        /// <remarks>This is provided by cosmos DB.</remarks>
-        string? Continuation { get; }
+        int NextPageNumber { get; }
     }
 }
